Use rails pipeline to load fonts and give placeholder profile images
Placeholder profile images are loaded as background images to work if the image url is found but does not work (current prod). This should be replaced with paperclip default images if they apply. Investigate this more when this project uses amazon s3 to store images (or a different solution). This pr also uses font-url instead of url when loading fonts which should find them in the rails asset pipeline.