Add banners to winner and fix stack preview bug
Stack preview would perform poorly with one card and also sometimes blow up way too big. This should now be fixed with a spacer for the single card (flex shrink was changing the height width, so just setting height didn't work) and a max height on the section for the stack preview.