Fix stack height bug in safari
For an element to have percentage height in safari its parent needs to have an explicitly stated height via the height attribute. letting flexbox handle the height of the element means safari uses the default height of the element instead of the listed percentage. this normally shouldn't be a problem as you can use flex to set its height, but i wanted auto height with the images so it broke hard.