Hide move bubble instead of removing it to maintain spacing
As a result of the last change to center the moves notifier the text would snap to center when a move was used. Now used moves are simply hidden via opacity meaning they still take up space. They shouldn't appear to any AT so it shouldn't be an issue for accessibility (they have no label or text content) and it maintains both the centering and the static spacing on the text